refactor(ai-service): Step 4 — move storage adapters to app/storage/
Local-filesystem persistence moves out of services/ into its
dedicated band. No network in this band; everything here owns
local-disk state owned by the AI service.

Moves:
  app/services/zarr_telemetry_store.py  →  app/storage/zarr/__init__.py
  app/services/model_cache_service.py   →  app/storage/cache/__init__.py

Importers updated (9 imports across 7 files):
  • app/services/{full_dataset_ml_service, imitate_expert_learning_service,
    backend_service, segment_classifier_service, segment_updater}.py
  • app/services/llm/telemetry_llm_orchestrator.py
  • ui/segment_tabs/shared.py

One follow-up: the zarr module used `parents[2]` to resolve the
project-root telemetry_zarr_store directory. After moving one level
deeper (services/X.py → storage/zarr/__init__.py), the right hop count
is `parents[3]`. Fixed inline.

Validated: `from app.storage.zarr import get_shared_zarr_store` resolves
to /app/telemetry_zarr_store (correct), /health returns 200, app.main
imports clean.
